feat(profile): implement lazy loading and split component
This commit implements lazy loading in the profile component and splits the component into two: Profile and ProfileDescriptions. Now, instead of using useMemo and returning an empty object, the same is done with lazy loading in the component and a loading bar is displayed inside the card instead of being blank.